% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{allele_match}
\alias{allele_match}
\title{For SNPs with equal chromosome and position, do the alleles match?}
\usage{
allele_match(query, reference)
}
\arguments{
\item{query}{a vector of (sorted) ldmap_snps SNPs}

\item{reference}{a vector of (sorted) ldmap_snps SNPs}
}
\description{
For SNPs with equal chromosome and position, do the alleles match?
}
