% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{alt_alleles}
\alias{alt_alleles}
\title{get ref alleles from a ldmap_snp}
\usage{
alt_alleles(ldmap_snp, as_ldmap_allele = TRUE)

alt_alleles(ldmap_snp, as_ldmap_allele = TRUE)
}
\arguments{
\item{ldmap_snp}{the vector of SNPs}
}
\description{
get ref alleles from a ldmap_snp

get ref alleles from a ldmap_snp
}
