% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{as_chromosome}
\alias{as_chromosome}
\title{Convert data to chromosome

Interprets the input as a chromosome
(which I model as integers/factors)}
\usage{
as_chromosome(x)
}
\arguments{
\item{x}{data to be converted to a format compatible with new_ldmap_snp or new_ldmap_region}
}
\value{
an integer vector (I may or may not make this a factor in the future)
}
\description{
Convert data to chromosome

Interprets the input as a chromosome
(which I model as integers/factors)
}
