% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{as_integer_ldmap_range}
\alias{as_integer_ldmap_range}
\title{coerce ldmap range to integer(64)}
\usage{
as_integer_ldmap_range(x)

as_integer_ldmap_range(x)
}
\arguments{
\item{x}{ldmap_range vec}
}
\description{
coerce ldmap range to integer(64)

coerce ldmap range to integer(64)
}
