% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_integer_ldmap_region}
\alias{as_integer_ldmap_region}
\title{coerce ldmap range to integer(64)}
\usage{
as_integer_ldmap_region(x)
}
\arguments{
\item{x}{ldmap_region vec}
}
\description{
coerce ldmap range to integer(64)
}
