% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_integer_ldmap_snp}
\alias{as_integer_ldmap_snp}
\title{coerce ldmap snp to integer(64)}
\usage{
as_integer_ldmap_snp(x)
}
\arguments{
\item{x}{ldmap_snp vec}
}
\description{
coerce ldmap snp to integer(64)
}
