% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{assign_region}
\alias{assign_region}
\title{Assign SNPs to LD blocks}
\usage{
assign_region(
  break_chr,
  break_start,
  break_stop,
  break_id = seq_along(break_chr),
  snp_chr,
  snp_pos,
  assign_all = T,
  max_size = -1L,
  min_size = 1L
)
}
\arguments{
\item{break_chr}{integer vector of chromosome assignment for each LD block}

\item{break_start}{integer vector of the start coordinate for each LD block}

\item{break_stop}{integer vector of the stop coordinate for each LD block}

\item{break_id}{vector with a unique id for each LD block}

\item{snp_chr}{integer vector giving the chromosome of every SNP}

\item{snp_pos}{integer vector giving the coordinate of each SNP}

\item{assign_all}{whether to throw an error if a SNP cannot be assigned to a block, or to assign it to block \code{NA}}

\item{max_size}{max number of snps to map to a region.  Each region will be split into sub-regions having no more than \code{max_size} elements,
unless \code{max_size} is less than or equal to 0, in which case there is no limit to the number of SNPs mapped to a region.}
}
\value{
modified \code{snp_df} dataframe with additional column \code{region_id} mapping snp to LD block
}
\description{
Assign SNPs to LD blocks
}
