% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_utils.R
\name{bed_region_cols}
\alias{bed_region_cols}
\title{Default columns for a \code{bed} (genomic regions) file}
\usage{
bed_region_cols(
  chrom = col_chromosome(),
  start = readr::col_integer(),
  end = readr::col_integer(),
  ...
)
}
\description{
Default columns for a \code{bed} (genomic regions) file
}
