% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{calc_AF}
\alias{calc_AF}
\title{Calculate allele frequency}
\usage{
calc_AF(x, na.rm = FALSE)
}
\arguments{
\item{x}{either a vector of type \code{ldmap_gt}, a list_of of type \code{ldmap_gt}, or a vector for which}

\item{na.rm}{whether or not to remove missing data}
}
\value{

}
\description{
Calculate allele frequency
}
