% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{calc_theta}
\alias{calc_theta}
\title{Calculate theta}
\usage{
calc_theta(m = 85)
}
\arguments{
\item{m}{genetic map sample size (85 for CEU from 1000 genomes estimate)}
}
\value{
scalar estimate of theta
}
\description{
Calculate theta
}
