% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{chromosomes}
\alias{chromosomes}
\title{get chroms from a ldmap_snp}
\usage{
chromosomes(struct_vec)

chromosomes(struct_vec)
}
\arguments{
\item{struct_vec}{the vector of SNPs (or ldmap_ranges)}
}
\description{
get chroms from a ldmap_snp

get chroms from a ldmap_snp
}
