% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{clear_alleles}
\alias{clear_alleles}
\title{Clear reference and alternate allele for ldmap_snp}
\usage{
clear_alleles(x)
}
\arguments{
\item{x}{a vector of ldmap_snps}
}
\value{
a vector of the length of x with chromosome and position set to \code{N}
}
\description{
Clear reference and alternate allele for ldmap_snp
}
