% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{col_chromosome}
\alias{col_chromosome}
\title{column specifier for chromosme}
\usage{
col_chromosome(prefix_chr = TRUE, ...)
}
\arguments{
\item{prefix_chr}{boolean indicating whether or not there is a \code{chr} prefix}

\item{...}{currently unused}
}
\value{
\code{col} specification (see \code{?readr::col})
}
\description{
column specifier for chromosme
}
