% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{common_range}
\alias{common_range}
\title{Overlapping interval}
\usage{
common_range(x, y)
}
\arguments{
\item{x}{vector of ldmap_snp or ldmap_range}

\item{y}{vector of ldmap_snp or ldmap_range}
}
\value{
vector of ldmap_range with overlap between x and y
}
\description{
Overlapping interval
}
