% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{common_region}
\alias{common_region}
\title{Overlapping interval}
\usage{
common_region(x, y)
}
\arguments{
\item{x}{vector of ldmap_snp or ldmap_region}

\item{y}{vector of ldmap_snp or ldmap_region}
}
\value{
vector of ldmap_region with overlap between x and y
}
\description{
Overlapping interval
}
