% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_utils.R
\name{compact_haplotype}
\alias{compact_haplotype}
\title{pack an array of 1s and 0s into integers}
\usage{
compact_haplotype(x, compact_64 = FALSE)
}
\arguments{
\item{x}{data consisting of  all \code{0}s and \code{1}s}

\item{compact_64}{whether to return an \code{int64} or not}
}
\value{
an integer array of \code{ceiling(length(x)/int_size)} where
\verb{int_size is 32 unless }compact_64\verb{is TRUE, in which case it's}64`
}
\description{
pack an array of 1s and 0s into integers
}
\author{
Nicholas Knoblauch
}
