% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{compact_ldmap_range}
\alias{compact_ldmap_range}
\title{convert  of a dataframe into a compact ldmap_range}
\usage{
compact_ldmap_range(
  df,
  chrom = "chrom",
  start = "start",
  end = "end",
  ldmap_range = "ldmap_range",
  remove = TRUE
)
}
\value{
copy of original datafame with new column \code{ldmap_range}
}
\description{
convert  of a dataframe into a compact ldmap_range
}
