% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{compact_ldmap_region}
\alias{compact_ldmap_region}
\title{convert  of a dataframe into a compact ldmap_region}
\usage{
compact_ldmap_region(
  df,
  chrom = "chrom",
  start = "start",
  end = "end",
  ldmap_region = "ldmap_region",
  remove = TRUE
)
}
\arguments{
\item{df}{dataframe}

\item{chrom}{chromosome column}

\item{start}{column name for beginning of region}

\item{end}{column name for end of region}

\item{ldmap_region}{name for new ldmap_region column}

\item{remove}{boolean indicating whether or not to remove aforementioned columns}
}
\value{
copy of original datafame with new column \code{ldmap_region}
}
\description{
convert  of a dataframe into a compact ldmap_region
}
