% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{compact_snp_struct}
\alias{compact_snp_struct}
\title{convert SNP columns of a dataframe into a compact ldmap_snp}
\usage{
compact_snp_struct(
  df,
  chrom = "chrom",
  pos = "pos",
  ref = "ref",
  alt = "alt",
  snp_struct = "snp_struct",
  remove = TRUE
)
}
\arguments{
\item{df}{dataframe}

\item{chrom}{chrom column}

\item{pos}{position column}

\item{ref}{reference allele column}

\item{alt}{alt allele column}

\item{snp_struct}{name of new ldmap_snp column}

\item{remove}{whether to remove these columns or keep them after compaction}
}
\value{
copy of original datafame with new column \code{snp_struct}
}
\description{
convert SNP columns of a dataframe into a compact ldmap_snp
}
