% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance.ldmap_snp.ldmap_snp}
\alias{distance.ldmap_snp.ldmap_snp}
\title{Quickly calculate distance between to ldmap_regions}
\usage{
\method{distance.ldmap_snp}{ldmap_snp}(query, target)
}
\arguments{
\item{query}{vector of query ldmap_snps}

\item{target}{vector of target ldmap_regions (must be sorted)}
}
\value{
a vector of integers with the length between the two ranges
}
\description{
Quickly calculate distance between to ldmap_regions
}
