% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_rr}
\alias{distance_rr}
\title{Quickly calculate distance between to ldmap_ranges}
\usage{
distance_rr(query, target)
}
\arguments{
\item{query}{vector of query ldmap_snps}

\item{target}{vector of target ldmap_ranges (must be sorted)}
}
\value{
a vector of integers with the length between the two ranges
}
\description{
Quickly calculate distance between to ldmap_ranges
}
