% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{dl_1kg_map}
\alias{dl_1kg_map}
\title{Download genetic map data from the 1000 genomes}
\usage{
dl_1kg_map(
  pop = "CEU",
  dest_dir = tempdir(),
  kg_url = "ftp://ftp-trace.ncbi.nlm.nih.gov/1000genomes/ftp/"
)
}
\arguments{
\item{pop}{the 1000 genomes population}

\item{dest_dir}{directory to download the data to}

\item{kg_url}{url with 1000 genomes data}
}
\value{
vector of paths files (see \code{read_1kg_maps})
}
\description{
Download genetic map data from the 1000 genomes
}
