% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{ends<-}
\alias{ends<-}
\title{Change end value for ldmap_range}
\usage{
ends(x) <- value
}
\arguments{
\item{x}{ldmap_range}

\item{value}{new end}
}
\value{
an ldmap_range with all starts at \code{value}
}
\description{
Change end value for ldmap_range
}
