% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{example_bigsnp}
\alias{example_bigsnp}
\title{bigsnpr dataset from EUR chromosome 22}
\usage{
example_bigsnp(td = tempdir())
}
\arguments{
\item{td}{directory to unzip into}
}
\value{
rds file path
}
\description{
example bigsnpr data
}
\author{
Nicholas Knoblauch
}
