% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{explode_ldmap_range}
\alias{explode_ldmap_range}
\title{convert ldmap_range back in to individual columns}
\usage{
explode_ldmap_range(df, ldmap_range = range_cols(df), remove = TRUE)
}
\arguments{
\item{df}{dataframe with at least one ldmap_snp_struct}

\item{remove}{boolean indicating whether \code{ldmap_range}
column should be removed}
}
\value{
dataframe with individual columns
}
\description{
convert ldmap_range back in to individual columns
}
