% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{explode_ldmap_region}
\alias{explode_ldmap_region}
\title{convert ldmap_region back in to individual columns}
\usage{
explode_ldmap_region(df, ldmap_region = region_cols(df), remove = TRUE)
}
\arguments{
\item{df}{dataframe with at least one ldmap_snp_struct}

\item{ldmap_region}{ldmap_region}

\item{remove}{boolean indicating whether \code{ldmap_region}
column should be removed}
}
\value{
dataframe with individual columns
}
\description{
convert ldmap_region back in to individual columns
}
