% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_alt}
\alias{extract_alt}
\title{Extract alternate allele from}
\usage{
extract_alt(ref, alleles_as_ambig)
}
\arguments{
\item{ref}{reference sequence (as obtained from a reference genome fasta file)}

\item{alleles_as_ambig}{IUPAC ambiguity codes representing alleles}
}
\description{
Extract alternate allele from
}
