% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_utils.R
\name{fam_cols}
\alias{fam_cols}
\title{Column specification for a plink bim file}
\usage{
fam_cols(
  fid = readr::col_character(),
  iid = readr::col_character(),
  father_id = readr::col_character(),
  mother_id = readr::col_character(),
  sex = readr::col_integer(),
  phenotype = readr::col_double(),
  ...
)
}
\arguments{
\item{...}{unused}

\item{chrom}{chromosome specifier}

\item{rsid}{snp id specifier}

\item{map}{genetic map specifier}

\item{pos}{position specifier}

\item{alt}{alt allele specifier}

\item{ref}{ref allele specifier}
}
\value{
\code{readr::cols} specification
}
\description{
Column specification for a plink bim file
}
