% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_alleles}
\alias{find_alleles}
\title{Find query SNP in a list of reference snps}
\usage{
find_alleles(
  query_chrom,
  query_pos,
  ref_chrom,
  ref_pos,
  query_chunk = as.integer(c()),
  ref_chunk = as.integer(c())
)
}
\arguments{
\item{query_chrom}{query chromosome}

\item{query_pos}{query position}

\item{ref_chrom}{reference chromosome}

\item{ref_pos}{reference position}

\item{query_chunk}{region assignment for query (optional)}

\item{ref_chunk}{region assignment for reference (optional)}
}
\value{
returns a vector with the position of the match, or NA if no match is found.
}
\description{
Find query SNP in a list of reference snps
}
