% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flip_alleles}
\alias{flip_alleles}
\title{Determine whether 2}
\usage{
flip_alleles(query_ref_alt, target_ref_alt)
}
\arguments{
\item{query_ref_alt}{is a ref/alt pair}

\item{target_ref_alt}{is another ref/alt pair}
}
\value{
returns a vector with 1 if the query matches the target, -1 if a flip is required, or 0 if they are incompatible;
Note that in the case that the reference and/or alternate allele are not single characters, they will only be checked for equality
}
\description{
Determine whether 2
}
