% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{format.ldmap_gt}
\alias{format.ldmap_gt}
\title{Formatting method for ldmap genotype}
\usage{
\method{format}{ldmap_gt}(x, ...)
}
\arguments{
\item{x}{ldmap_gt}

\item{...}{unused}
}
\value{

}
\description{
Formatting method for ldmap genotype
}
