% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.R
\name{format.ldmap_ht}
\alias{format.ldmap_ht}
\title{Formatting method for ldmap allele}
\usage{
\method{format}{ldmap_ht}(x, ...)
}
\arguments{
\item{x}{ldmap_ht}

\item{...}{unused}
}
\value{

}
\description{
Formatting method for ldmap allele
}
