% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{format.ldmap_range}
\alias{format.ldmap_range}
\title{Formatting method for ldmap ranges}
\usage{
\method{format}{ldmap_range}(x, ...)
}
\arguments{
\item{x}{ldmap_range}

\item{...}{}
}
\value{

}
\description{
Formatting method for ldmap ranges
}
