% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{format.ldmap_region}
\alias{format.ldmap_region}
\title{Formatting method for ldmap regions}
\usage{
\method{format}{ldmap_region}(x, ...)
}
\arguments{
\item{x}{ldmap_region}

\item{...}{}
}
\value{

}
\description{
Formatting method for ldmap regions
}
