% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{format.ldmap_snp}
\alias{format.ldmap_snp}
\title{Formatting method for ldmap snps}
\usage{
\method{format}{ldmap_snp}(x, ...)
}
\arguments{
\item{x}{ldmap_snp}

\item{...}{unused}
}
\value{

}
\description{
Formatting method for ldmap snps
}
