% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{format_ldmap_allele}
\alias{format_ldmap_allele}
\title{get ref alleles from a ldmap_snp}
\usage{
format_ldmap_allele(x)
}
\arguments{
\item{x}{ldmap allele vec}
}
\description{
get ref alleles from a ldmap_snp
}
