% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{format_ldmap_range}
\alias{format_ldmap_range}
\title{formatting of ldmap_ranges}
\usage{
format_ldmap_range(x)

format_ldmap_range(x)
}
\arguments{
\item{x}{an ldmap_range}
}
\description{
formatting of ldmap_ranges

formatting of ldmap_ranges
}
