% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{format_ldmap_region}
\alias{format_ldmap_region}
\title{formatting of ldmap_regions}
\usage{
format_ldmap_region(x)
}
\arguments{
\item{x}{an ldmap_region}
}
\description{
formatting of ldmap_regions
}
