% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{format_ldmap_snp}
\alias{format_ldmap_snp}
\title{Formatting method for ldmap snps}
\usage{
format_ldmap_snp(x)
}
\arguments{
\item{x}{a vector of ldmap_snps}
}
\description{
Formatting method for ldmap snps
}
