% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{group_regions}
\alias{group_regions}
\title{Assign ldmap_regions into non-overlapping groups (}
\usage{
group_regions(query)
}
\arguments{
\item{query}{vector of type ldmap_region}
}
\description{
Assign ldmap_regions into non-overlapping groups (
}
