% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gt2matrix}
\alias{gt2matrix}
\title{Convert list of ldmap_gt to numeric matrix}
\usage{
gt2matrix(x)
}
\arguments{
\item{x}{a vector of type ldmap_gt}
}
\description{
Convert list of ldmap_gt to numeric matrix
}
