% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{int2rsid}
\alias{int2rsid}
\title{quickly convert a vector of rsids from integer to string representation}
\usage{
int2rsid(x)
}
\arguments{
\item{x}{vector of rsids}
}
\value{

}
\description{
quickly convert a vector of rsids from integer to string representation
}
