% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interpolate_genetic_map}
\alias{interpolate_genetic_map}
\title{Linear interpolation of genetic map values}
\usage{
interpolate_genetic_map(
  map,
  map_pos,
  target_pos,
  strict = TRUE,
  progress = FALSE
)
}
\arguments{
\item{map}{is a length \code{p} vector of cumulative genetic map values. \code{map} must be \emph{strictly} \emph{sorted}}

\item{map_pos}{is a length \code{p} vector of genome coordinates corresponding to the reference genetic map. \code{map_pos} must be \emph{strictly} \emph{sorted}}

\item{target_pos}{is a vector of coordinates to interpolate}

\item{strict}{a boolean indicating whether to strictly interpolate}

\item{progress}{a boolean indicating whether to indicate progress with a progress bar}
}
\description{
Linear interpolation of genetic map values
}
