% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{is_ldmap}
\alias{is_ldmap}
\title{Predicate testing for whether an object is of a type in \code{ldmap}}
\usage{
is_ldmap(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
scalar logical
}
\description{
Predicate testing for whether an object is of a type in \code{ldmap}
}
