% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{is_ldmap_range}
\alias{is_ldmap_range}
\title{Whether or not your vector is a ldmap_range}
\usage{
is_ldmap_range(x)

is_ldmap_range(x)
}
\arguments{
\item{x}{vector}
}
\value{
length one boolean

length one boolean
}
\description{
Whether or not your vector is a ldmap_range

Whether or not your vector is a ldmap_snp
}
\examples{
#FALSE
is_ldmap_range(letters)
#TRUE
is_ldmap_range(new_ldmap_range(1:22,1:100,200:300))
#FALSE
is_ldmap_snp(letters)
#TRUE
is_ldmap_snp(new_ldmap_snp(1:22,1:100))
}
