% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{is_ldmap_region}
\alias{is_ldmap_region}
\title{Whether or not your vector is a ldmap_region}
\usage{
is_ldmap_region(x)
}
\arguments{
\item{x}{vector}
}
\value{
length one boolean
}
\description{
Whether or not your vector is a ldmap_region
}
\examples{
#FALSE
is_ldmap_region(letters)
#TRUE
is_ldmap_region(new_ldmap_region(1:22,1:100,200:300))
}
