% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{is_range_in_range}
\alias{is_range_in_range}
\title{Check for membership of a range in another range}
\usage{
is_range_in_range(query, target, allow_overlap = FALSE)
}
\arguments{
\item{query}{vector of ldmap_ranges}

\item{target}{vector of ldmap_ranges}

\item{allow_overlap}{whether partial overlap counts as membership}
}
\value{
boolean vector of length(query)
}
\description{
Check for membership of a range in another range
}
