% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{is_region_in_region}
\alias{is_region_in_region}
\title{Check for membership of a region in another region}
\usage{
is_region_in_region(query, target, allow_overlap = FALSE)
}
\arguments{
\item{query}{vector of ldmap_regions}

\item{target}{vector of ldmap_regions}

\item{allow_overlap}{whether partial overlap counts as membership}
}
\value{
boolean vector of length(query)
}
\description{
Check for membership of a region in another region
}
