% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{is_reversed}
\alias{is_reversed}
\title{Are ref and alt reversed?}
\usage{
is_reversed(x)
}
\arguments{
\item{x}{a factor as returned from \code{allele_match}}
}
\value{
a logical vector of length \code{length(x)}
}
\description{
Are ref and alt reversed?
}
