% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{is_snp_in_range}
\alias{is_snp_in_range}
\title{Check for membership of a SNP in a region/range}
\usage{
is_snp_in_range(ldmap_snp, ldmap_range)
}
\arguments{
\item{ldmap_snp}{vector of ldmap_snps}

\item{ldmap_range}{vector of ldmap_ranges}
}
\value{
boolean vector of length(length(ldmap_snp))
}
\description{
Check for membership of a SNP in a region/range
}
