% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{is_snp_in_region}
\alias{is_snp_in_region}
\title{Check for membership of a SNP in a region/region}
\usage{
is_snp_in_region(ldmap_snp, ldmap_region)
}
\arguments{
\item{ldmap_snp}{vector of ldmap_snps}

\item{ldmap_region}{vector of ldmap_regions}
}
\value{
boolean vector of length(length(ldmap_snp))
}
\description{
Check for membership of a SNP in a region/region
}
