% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{join_snp}
\alias{join_snp}
\title{Find SNPs that match (and find out what to do with them)}
\usage{
join_snp(query, reference, rsid = as.integer(c()))

join_snp(query, reference, rsid = as.integer(c()))
}
\arguments{
\item{query}{a vector of (sorted) ldmap_snps SNPs}

\item{reference}{a vector of (sorted) ldmap_snps SNPs}
}
\description{
Find SNPs that match (and find out what to do with them)

Find SNPs that match (and find out what to do with them)
}
