% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldmap-package.R
\docType{package}
\name{ldmap-package}
\alias{ldmap}
\alias{ldmap-package}
\title{ldmap: Manipulate SNP coordinates and Interpolate Genetic Map Data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CreRecombinase/ldmap}
  \item Report bugs at \url{https://github.com/CreRecombinase/ldmap/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
