% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{ldmap_range}
\alias{ldmap_range}
\title{New ldmap_range}
\usage{
ldmap_range(x = double())
}
\arguments{
\item{x}{vector}
}
\value{
ldmap_range
}
\description{
New ldmap_range
}
\examples{
#under the hood the data is a double
as.numeric(ldetect_EUR[1])
#which means we can convert it back
ldmap_range(as.numeric(ldetect_EUR[1]))
}
