% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{ldmap_range_2_data_frame}
\alias{ldmap_range_2_data_frame}
\title{convert ldmap_range to dataframe}
\usage{
ldmap_range_2_data_frame(ldmap_range)

ldmap_range_2_data_frame(ldmap_range)
}
\arguments{
\item{ldmap_range}{an ldmap_range}
}
\description{
convert ldmap_range to dataframe

convert ldmap_range to dataframe
}
