% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{ldmap_region}
\alias{ldmap_region}
\title{New ldmap_region}
\usage{
ldmap_region(x = double())
}
\arguments{
\item{x}{vector}
}
\value{
ldmap_region
}
\description{
New ldmap_region
}
\examples{
#under the hood the data is a double
as.numeric(ldetect_EUR[1])
#which means we can convert it back
ldmap_region(as.numeric(ldetect_EUR[1]))
}
