% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldmap_region_2_data_frame}
\alias{ldmap_region_2_data_frame}
\title{convert ldmap_region to dataframe}
\usage{
ldmap_region_2_data_frame(ldmap_region)
}
\arguments{
\item{ldmap_region}{an ldmap_region}
}
\description{
convert ldmap_region to dataframe
}
