% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldmap_snp_2_dataframe}
\alias{ldmap_snp_2_dataframe}
\title{Formatting method for ldmap snps}
\usage{
ldmap_snp_2_dataframe(struct_vec)
}
\arguments{
\item{struct_vec}{the vector of SNPs}
}
\description{
Formatting method for ldmap snps
}
