% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{left_snp_join}
\alias{left_snp_join}
\title{Join dataframes with ldmap_snps}
\usage{
left_snp_join(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  snp_struct_col = snp_cols(gwas_df),
  flip_sign_col = c("beta"),
  rename_col = TRUE,
  remove_missing = FALSE
)
}
\arguments{
\item{x}{dataframe with at least one ldmap_snp column to join on}

\item{y}{dataframe with at least one ldmap_snp column}

\item{by}{column name of snp column}

\item{suffix}{suffix to add}

\item{snp_struct_col}{column name of ldmap_snp in \code{gwas_df}}

\item{flip_sign_col}{columns that should be flipped when the query SNP is flipped against the target}

\item{rename_col}{whether to rename column}

\item{remove_missing}{whether to remove missing data}

\item{remove_ambig}{whether to remove strand ambiguous query SNPs prior to trying to match target}
}
\value{

}
\description{
Join dataframes with ldmap_snps
}
