% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{make_ambig}
\alias{make_ambig}
\title{Create ambiguity codes from two alleles}
\usage{
make_ambig(A1, A2)

make_ambig(A1, A2)
}
\arguments{
\item{A1}{allele 1}

\item{A2}{allele 2}
}
\description{
Create ambiguity codes from two alleles

Create ambiguity codes from two alleles
}
