% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_ranges_snps}
\alias{match_ranges_snps}
\title{Assign SNPs to ranges}
\usage{
match_ranges_snps(df, ldmap_range, snp_col = "snp_struct")
}
\arguments{
\item{ldmap_range}{vector of potentially overlapping ldmap_ranges (must be sorted)}

\item{ldmap_snp}{vector of ldmap_snps (must be sorted)}
}
\value{
a list of integer vectors giving the ranges to which each SNP belongs
}
\description{
Assign SNPs to ranges
}
