% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{merge_ldmap_ranges}
\alias{merge_ldmap_ranges}
\title{Merge two ldmap_range vectors}
\usage{
merge_ldmap_ranges(x, y)

merge_ldmap_ranges(x, y)
}
\arguments{
\item{x}{a (preferably sorted) ldmap_range vector (\code{length(x)} is referred to  as  \code{p})}

\item{y}{a (preferably sorted) ldmap_range vector (\code{length(y)} is referred to  as  \code{q})}
}
\value{
a sorted vector ldmap_ranges of length at most \code{p+q} and at least \code{max(p,q)} representing the union of the two sets of ranges

a sorted vector ldmap_ranges of length at most \code{p+q} and at least \code{max(p,q)} representing the union of the two sets of ranges
}
\description{
Merge two ldmap_range vectors

Merge two ldmap_range vectors
}
