% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_ldmap_regions}
\alias{merge_ldmap_regions}
\title{Merge two ldmap_region vectors}
\usage{
merge_ldmap_regions(x, y)
}
\arguments{
\item{x}{a (preferably sorted) ldmap_region vector (\code{length(x)} is referred to  as  \code{p})}

\item{y}{a (preferably sorted) ldmap_region vector (\code{length(y)} is referred to  as  \code{q})}
}
\value{
a sorted vector ldmap_regions of length at most \code{p+q} and at least \code{max(p,q)} representing the union of the two sets of ranges
}
\description{
Merge two ldmap_region vectors
}
