% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{nearest_snp_range}
\alias{nearest_snp_range}
\title{Assign ranges to nearest ranges}
\usage{
nearest_snp_range(query, target)

nearest_snp_range(query, target)
}
\arguments{
\item{query}{vector of query ldmap_snps}

\item{target}{vector of target ldmap_ranges (must be sorted)}

\item{use_begin}{logical scalar indicating whether to use the start of the target range(TRUE), the end(FALSE), or them min distance of the two (NA_LOGICAL)}
}
\value{
a vector of integers of length \code{length(ldmap_range_query)} with the index of the \code{ldmap_range_target}

a vector of integers of length \code{length(ldmap_range_query)} with the index of the \code{ldmap_range_target} (or NA_INTEGER if there is no overlap in the set of chromosomes)
}
\description{
Assign ranges to nearest ranges

Assign ranges to nearest ranges
}
