% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{new_ldmap_allele}
\alias{new_ldmap_allele}
\title{Convert allele info to ldmap_alleles}
\usage{
new_ldmap_allele(allele = as.integer(c()))

new_ldmap_allele(allele = as.integer(c()))
}
\arguments{
\item{allele}{vector of alleles, coded either as character, integer or raw}
}
\description{
Convert allele info to ldmap_alleles

Convert allele info to ldmap_alleles
}
