% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{new_ldmap_gt}
\alias{new_ldmap_gt}
\title{Create a vector of plink genotypes from a raw byte vector}
\usage{
new_ldmap_gt(x, N = length(x) * 4)

new_ldmap_gt(x, N = length(x) * 4)
}
\arguments{
\item{x}{vector of bytes of type raw}

\item{N}{number of samples in the data(otherwise assumed it is length(x)*4}
}
\value{
a vector of ldmap_gt

a vector of ldmap_gt
}
\description{
Create a vector of plink genotypes from a raw byte vector

Create a vector of plink genotypes from a raw byte vector
}
