% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_ldmap_ht}
\alias{new_ldmap_ht}
\title{New vector of haplotypes}
\usage{
new_ldmap_ht(x = as.integer(c()))
}
\arguments{
\item{x}{data vector}

\item{N}{sample size}
}
\value{
vector of (packed, dense) haplotypes
}
\description{
New vector of haplotypes
}
