% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_ldmap_range}
\alias{new_ldmap_range}
\title{Creation of new ldmap_ranges}
\usage{
new_ldmap_range(
  chrom = as.integer(c()),
  start = as.integer(c()),
  end = as.integer(c())
)
}
\arguments{
\item{chrom}{an integer vector of chromosomes}

\item{start}{an integer vector of start positions}

\item{stop}{an integer vector of stop positions}
}
\description{
Creation of new ldmap_ranges
}
