% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{new_ldmap_region}
\alias{new_ldmap_region}
\title{Construct new ldmap_region}
\usage{
new_ldmap_region(chrom = integer(), start = integer(), end = integer())
}
\arguments{
\item{chrom}{chromosome}

\item{start}{start}

\item{end}{end}
}
\value{

}
\description{
Construct new ldmap_region
}
\author{

}
