% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_ldmap_snp_impl}
\alias{new_ldmap_snp_impl}
\title{Creation of new ldmap_snps}
\usage{
new_ldmap_snp_impl(chrom, pos, ref, alt, NA2N = FALSE)
}
\arguments{
\item{chrom}{an integer vector of chromosomes}

\item{pos}{a double vector of positions}

\item{ref}{an optional vector of reference allele (see \code{?new_ldmap_allele})}

\item{alt}{an optional vector of alternate allele (see \code{?new_ldmap_allele})}

\item{NA2N}{an optional boolean specifying whether missing/NA alleles should be treated as "N"}
}
\description{
Creation of new ldmap_snps
}
