% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{order.ldmap_snp}
\alias{order.ldmap_snp}
\title{sorting method for ldmap snps}
\usage{
\method{order}{ldmap_snp}(struct_vec)

\method{order}{ldmap_snp}(struct_vec)
}
\arguments{
\item{struct_vec}{the vector of SNPs}
}
\description{
sorting method for ldmap snps

sorting method for ldmap snps
}
