% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsnp.R
\name{panel_ld}
\alias{panel_ld}
\title{estimate LD from a reference panel}
\usage{
panel_ld(reference_file, LDshrink = TRUE)
}
\arguments{
\item{LDshrink}{boolean for whether to use ldshrink}

\item{reference_files}{a vector of 22 bigsnpr files}
}
\value{
LD matrix
}
\description{
estimate LD from a reference panel
}
\author{
Nicholas Knoblauch
}
