% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_ldmap_region}
\alias{parse_ldmap_region}
\title{Creation of new ldmap_region vector from character}
\usage{
parse_ldmap_region(input)
}
\arguments{
\item{input}{a character vector}
}
\description{
Creation of new ldmap_region vector from character
}
