% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_ldmap_SNP}
\alias{parse_ldmap_SNP}
\title{Creation of new ldmap_snp vector from character}
\usage{
parse_ldmap_SNP(input)
}
\arguments{
\item{input}{a character vector}
}
\description{
Creation of new ldmap_snp vector from character
}
