% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{positions<-}
\alias{positions<-}
\title{Change position for a vector of ldmap_snps}
\usage{
positions(x) <- value
}
\arguments{
\item{x}{ldmap_snp}

\item{value}{new position}
}
\value{
an ldmap_snp vector with all positions at \code{value}
}
\description{
Change position for a vector of ldmap_snps
}
