% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{positions}
\alias{positions}
\title{get positions from a ldmap_snp}
\usage{
positions(ldmap_snp)

positions(ldmap_snp)
}
\arguments{
\item{ldmap_snp}{the vector of SNPs}
}
\description{
get positions from a ldmap_snp

get positions from a ldmap_snp
}
