% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_utils.R
\name{query_reference_snpset}
\alias{query_reference_snpset}
\title{Query a a reference panel for a set of SNPs and see if any of them need a sign flip}
\usage{
query_reference_snpset(
  query_chr,
  query_pos,
  query_allele,
  query_block = NULL,
  panel_chr,
  panel_pos,
  panel_allele,
  panel_block = NULL
)
}
\arguments{
\item{query_chr}{chromosome for query SNPs coded as integer or character (coding must match \code{panel_chr})}

\item{query_pos}{position of query SNPs}

\item{query_allele}{allele for query SNPs}

\item{query_block}{optional vector specifiyng a chunking scheme (default is NULL)}

\item{panel_chr}{chromosome for the reference SNPs coded as integer or character (coding must match \code{query_chr})}

\item{panel_pos}{position of reference SNPs}

\item{panel_allele}{allele for reference SNPs}

\item{panel_block}{optional vector specifying a chunking scheme (default is NULL)}
}
\value{
an \code{integer} vector of length equal to \code{length(query_chr)} where a \code{1} indicates that the query position
is in the reference panel, a \code{0} indicates that it is not, and a \code{-1} indicates a match
, but that a flip is necessary
}
\description{
Query a a reference panel for a set of SNPs and see if any of them need a sign flip
}
\examples{

#read in the example data
example_bim <- fs::path_package(package = "ldshrink","test_data/reference_genotype.bim")
panel_df <- read.table(example_bim,sep="\\t",col.names=c("chr","snp","map","pos","ref","alt"))
#create "allele" column
panel_allele <- paste0(panel_df$ref,",",panel_df$alt)
stopifnot(all(nchar(panel_allele)==3))
}
