% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{range_cols}
\alias{range_cols}
\title{Return the column(s) containing ldmap_ranges}
\usage{
range_cols(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a character vector with the names \code{ldmap_range} columns
}
\description{
Return the column(s) containing ldmap_ranges
}
\author{
Nicholas Knoblauch
}
