% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{range_in_range}
\alias{range_in_range}
\title{Assign ranges to ranges}
\usage{
range_in_range(ldmap_range_query, ldmap_range_target)
}
\arguments{
\item{ldmap_range_query}{vector of ldmap_ranges}

\item{ldmap_range_target}{vector of \emph{non-overlapping} ldmap_ranges (must be sorted)}
}
\value{
a vector of integers of length \code{length(ldmap_range_query)} with the index of the \code{ldmap_range_target}
}
\description{
Assign ranges to ranges
}
