% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{rank.ldmap_snp}
\alias{rank.ldmap_snp}
\title{ranking method for ldmap snps}
\usage{
\method{rank}{ldmap_snp}(struct_vec)

\method{rank}{ldmap_snp}(struct_vec)
}
\arguments{
\item{struct_vec}{the vector of SNPs}
}
\description{
ranking method for ldmap snps

ranking method for ldmap snps
}
