% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{read_fam}
\alias{read_fam}
\title{read plink fam file}
\usage{
read_fam(x, cols = fam_cols())
}
\arguments{
\item{x}{path to file}

\item{cols}{fam columns (a readr cols object)}
}
\value{
a dataframe with fam info
}
\description{
read plink fam file
}
