% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_fasta}
\alias{read_fasta}
\title{Read the contents of a fasta file as a vector of ldmap_alleles}
\usage{
read_fasta(fastafile, subset = NULL, line_length = 50L)
}
\arguments{
\item{fastafile}{path to a fasta file (can be gzipped)}

\item{subset}{positions to subset in the file.  Default is the whole sequence}

\item{line_length}{the number of nucleotide characters per line (default is 50)}
}
\value{
a vector ldmap_allele with the sequence contents of the file
}
\description{
Read the contents of a fasta file as a vector of ldmap_alleles
}
