% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap_utils.R
\name{read_hap_legend}
\alias{read_hap_legend}
\title{read haplegend}
\usage{
read_hap_legend(x, chr = guess_chr(x), read_fun = readr::read_delim)
}
\arguments{
\item{x}{path to haplegend file}
}
\value{
tibble with 1 ldmap_snp column
}
\description{
read haplegend
}
