% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.R
\name{read_hap_samples}
\alias{read_hap_samples}
\title{read haploid sample info##'}
\usage{
read_hap_samples(x)
}
\arguments{
\item{x}{\code{.samples} file}
}
\value{
dataframe with sample info
}
\description{
read haploid sample info##'
}
