% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.R
\name{read_hap_snps}
\alias{read_hap_snps}
\title{read regions of}
\usage{
read_hap_snps(hap_dir, rsid, hap_file_pattern = "EUR.chr{chrom}_1kg_geno")
}
\arguments{
\item{hap_dir}{directory with haplotype data}

\item{hap_file_pattern}{\code{glue}-compatible file pattern for haplotype files,
when '.legend.gz' is added as a suffix, and \code{chrom} is substituted, should resolve to a
haplotype legend file.  Similarly when \code{.hap.gz} is added as a suffix should resolve to a
haplotype file}

\item{ldmr}{ldmap_region(s) to read}
}
\value{
a dataframe or list of dataframes (see \code{bind_rows})
}
\description{
read regions of
}
