% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{read_plink}
\alias{read_plink}
\title{Read plink bed file into a (relatively) tidy format}
\usage{
read_plink(f, return_fam = FALSE)
}
\arguments{
\item{f}{path to plink \code{.bed} file}
}
\value{

}
\description{
Read plink bed file into a (relatively) tidy format
}
