% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_snp_rsid_h5}
\alias{read_snp_rsid_h5}
\title{read a snp dataframe stored in HDF5 with a snp index}
\usage{
read_snp_rsid_h5(
  h5file,
  snplist,
  datapath = "snp",
  snp_id = FALSE,
  rsidpath = "rsid",
  allow_missing = FALSE,
  ...
)
}
\arguments{
\item{h5file}{hdf5 file}

\item{snplist}{list of rsids to read (alternatively a list of ldmap_snps}

\item{datapath}{hdf5 group that contains the dataframe}

\item{snp_id}{boolean for whether rownames are returned}

\item{rsidpath}{dataset name of the snp column to match}

\item{allow_missing}{whether missing data is allowed}
}
\value{

}
\description{
read a snp dataframe stored in HDF5 with a snp index
}
