% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{ref_alleles<-}
\alias{ref_alleles<-}
\title{Change ref_allele for a vector of ldmap_snps}
\usage{
ref_alleles(x) <- value
}
\arguments{
\item{x}{ldmap_snp}

\item{value}{new ref_allele}
}
\value{
an ldmap_snp vector with ref_alleles equal to \code{value}
}
\description{
Change ref_allele for a vector of ldmap_snps
}
