% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsnp.R
\name{reference_panel_ld}
\alias{reference_panel_ld}
\title{estimate reference panel LD}
\usage{
reference_panel_ld(
  gwas_df,
  reference_file,
  LDshrink = TRUE,
  drop_missing = TRUE
)
}
\arguments{
\item{gwas_df}{dataframe with a ldmap_snp column (and optionally an effect-size column (beta-hat)}

\item{reference_file}{\code{rds} file with \code{bigsnp} data}

\item{LDshrink}{boolean whether to use LDshrink}

\item{drop_missing}{whether to drop missing data or throw an error}
}
\value{
correlation matrix
}
\description{
estimate reference panel LD
}
\author{
Nicholas Knoblauch
}
