% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{region_cols}
\alias{region_cols}
\title{Return the column(s) containing ldmap_regions}
\usage{
region_cols(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a character vector with the names \code{ldmap_region} columns
}
\description{
Return the column(s) containing ldmap_regions
}
\author{
Nicholas Knoblauch
}
