% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{region_in_region}
\alias{region_in_region}
\title{Assign ranges to ranges}
\usage{
region_in_region(
  ldmap_region_query,
  ldmap_region_target,
  allow_overlap = FALSE
)
}
\arguments{
\item{ldmap_region_query}{vector of ldmap_regions}

\item{ldmap_region_target}{vector of \emph{non-overlapping} ldmap_regions (must be sorted)}

\item{allow_overlap}{is it alright if a query is only partially inside the target?}
}
\value{
a vector of integers of length \code{length(ldmap_region_query)} with the index of the \code{ldmap_region_target}
}
\description{
Assign ranges to ranges
}
