% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region_overlap_region}
\alias{region_overlap_region}
\title{Assign ranges to ranges}
\usage{
region_overlap_region(ldmap_region_query, ldmap_region_target)
}
\arguments{
\item{ldmap_region_query}{vector of ldmap_regions}

\item{ldmap_region_target}{vector of \emph{non-overlapping} ldmap_regions (must be sorted)}
}
\value{
a vector of integers of length \code{length(ldmap_region_query)} with the index of the \code{ldmap_region_target}
}
\description{
Assign ranges to ranges
}
