% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{region_overlap_snp}
\alias{region_overlap_snp}
\title{Assign SNPs to ranges}
\usage{
region_overlap_snp(ldmap_region, ldmap_snp)
}
\arguments{
\item{ldmap_region}{vector of non-overlapping ldmap_regions (must be sorted)}

\item{ldmap_snp}{vector of ldmap_snps (must be sorted)}
}
\value{
a vector of integers of length \code{length(ldmap_snp)} with the index of the \code{ldmap_region}
}
\description{
Assign SNPs to ranges
}
