% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rregion}
\alias{rregion}
\title{Generate random genomic regions}
\usage{
rregion(n, chroms = 1L:23L, sort = TRUE)
}
\arguments{
\item{n}{number of ranges to generate}

\item{chroms}{chromosomes from which to sample}

\item{sort}{whether to sort the output}
}
\value{
a vector of length \code{n} of (optionally sorted ldmap_ranges)
}
\description{
Generate random genomic regions
}
\examples{
#chromosomes are assumed to be hg19
head(rregion(100,chroms=1L))
}
