% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{rsid2int}
\alias{rsid2int}
\title{quickly convert a vector of rsids from string to integer representation}
\usage{
rsid2int(x)
}
\arguments{
\item{x}{vector of rsids}
}
\value{

}
\description{
quickly convert a vector of rsids from string to integer representation
}
