% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rsnp}
\alias{rsnp}
\title{Draw SNPs from random positions in the genome}
\usage{
rsnp(n, chroms = 1L:23L, sort = TRUE, replace = FALSE, ...)
}
\arguments{
\item{n}{number of SNPs to simulate}

\item{chroms}{chromosomes from which to draw}

\item{sort}{whether to sort output}

\item{...}{currently unused}
}
\value{

}
\description{
Draw SNPs from random positions in the genome
}
\author{
Nicholas Knoblauch
}
