% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rsnp_region}
\alias{rsnp_region}
\title{Draw random snps from ldmap regions}
\usage{
rsnp_region(ldmap_region, n, sort = TRUE, replace = FALSE)
}
\arguments{
\item{ldmap_region}{ldmap region to draw samples from}

\item{n}{number of draws, recycled to match ldmap_region}

\item{sort}{whether to sort the result}
}
\value{
vector of ldmap_snps
}
\description{
Draw random snps from ldmap regions
}
\examples{
#generate one snp per ldetect block
data(ldetect_EUR)
snp_per_block <- rsnp_region(ldetect_EUR,n=1)
##generate three snps from one ldetect block
snps <- rsnp_region(ldetect_EUR[4],n=3)
##generate 10 snps in total, 3 from one ldetect block
##four from the second and 3 from the third
snps <- rsnp_region(ldetect_EUR[4:6],n=c(3,4))
}
