% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{set_ends}
\alias{set_ends}
\title{Change end position of a region}
\usage{
set_ends(x, to)
}
\arguments{
\item{x}{source vector of region}

\item{to}{new position value for region}
}
\value{
a vector of length equal to \code{x} with all(ends(x)==to)
}
\description{
Change end position of a region
}
\examples{
#generate random snps and then set chromosmes to 1:10 and all positions to 10
x <- ldmap_snp(double(10)) \%>\% set_chromosomes(1:10) \%>\% set_positions(10)
all(positions(x)==10)
}
