% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_ld_region}
\alias{set_ld_region}
\title{Assign snps to regions of the genome, breaking up regions based on number of SNPs}
\usage{
set_ld_region(
  ld_chr,
  ld_start,
  ld_stop,
  ld_region_id,
  chr,
  pos,
  max_size = 0L,
  min_size = 1L,
  assign_all = TRUE
)
}
\arguments{
\item{ld_chr}{vector of chromosomes    of per region}

\item{ld_start}{vector of start positions for each region}

\item{ld_stop}{vector of end positions for each region}
}
\value{
returns a vector with 1 if the query matches the target, -1 if a flip is required, or 0 if they are incompatible;
}
\description{
Assign snps to regions of the genome, breaking up regions based on number of SNPs
}
