% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{set_positions}
\alias{set_positions}
\title{Change position of a SNP}
\usage{
set_positions(x, to)
}
\arguments{
\item{x}{source vector of SNPs}

\item{to}{new position value for SNPs}
}
\value{
a vector of length equal to \code{x} with all(positions(x)==to)
}
\description{
Change position of a SNP
}
\examples{
x <- ldmap_snp(double(10)) \%>\% set_chromosomes(1:10) \%>\% set_positions(10)
all(positions(x)==10)
}
