% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{snp2raw}
\alias{snp2raw}
\title{Determine whether 2 alleles are compatible}
\usage{
snp2raw(input_matrix)
}
\arguments{
\item{query_ref_alt}{is a ref/alt pair}

\item{target_ref_alt}{is another ref/alt pair}
}
\value{
raw matrix
}
\description{
Determine whether 2 alleles are compatible
}
