% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{snp_cols}
\alias{snp_cols}
\title{Return the column(s) containing ldmap_snps}
\usage{
snp_cols(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a character vector with the names \code{ldmap_snp} columns
}
\description{
Return the column(s) containing ldmap_snps
}
\author{
Nicholas Knoblauch
}
