% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{snp_in_range}
\alias{snp_in_range}
\title{formatting of ldmap_ranges}
\usage{
snp_in_range(ldmap_snp, ldmap_range)
}
\arguments{
\item{ldmap_snp}{vector of ldmap_snps}

\item{ldmap_range}{vector of ldmap_snps}
}
\description{
formatting of ldmap_ranges
}
