% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{snp_in_regions}
\alias{snp_in_regions}
\title{Assign SNPs to ranges}
\usage{
snp_in_regions(ldmap_snp, ldmap_regions)
}
\arguments{
\item{ldmap_snp}{vector of ldmap_snps (must be sorted)}

\item{ldmap_region}{vector of non-overlapping ldmap_regions (must be sorted)}
}
\value{
a vector of integers of length \code{length(ldmap_snp)} with the index of the \code{ldmap_region}
}
\description{
Assign SNPs to ranges
}
