% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/snp.R
\name{snp_overlap_region}
\alias{snp_overlap_region}
\title{Check SNPs for positional equality}
\usage{
snp_overlap_region(ldmap_snp, ldmap_region)

snp_overlap_region(ldmap_snp, ldmap_region)
}
\arguments{
\item{ldmap_snp}{vector of ldmap_snps (must be sorted)}

\item{ldmap_region}{vector of non-overlapping ldmap_regions (must be sorted)}

\item{x}{vector of query ldmap_snps}

\item{y}{vector of target ldmap_snps}
}
\value{
a vector of integers of length \code{x} that indexes into \code{y}

a vector of integers of length \code{length(ldmap_snp)} with the index of the \code{ldmap_region}
}
\description{
Check SNPs for positional equality

Assign SNPs to ranges
}
