% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sorted_snp_df}
\alias{sorted_snp_df}
\title{Find out of vector of SNPs is sorted}
\usage{
sorted_snp_df(chr, pos)
}
\arguments{
\item{chr}{vector of chromosomes    of per region}

\item{pos}{vector of start positions for each region}
}
\value{
returns true if the vector is sorted
}
\description{
Find out of vector of SNPs is sorted
}
