% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_by_overlap}
\alias{split_by_overlap}
\title{Split a vector ldmap_regions into non-overlapping groups}
\usage{
split_by_overlap(x)
}
\arguments{
\item{x}{a vector of ldmap_snps}
}
\description{
Split a vector ldmap_regions into non-overlapping groups
}
