% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{split_ldmap_range_overlap}
\alias{split_ldmap_range_overlap}
\title{Take a vector of (preferably) sorted and (possibly) overlapping ldmap_ranges and create a new range of (sorted) non-overlapping ldmap_ranges}
\usage{
split_ldmap_range_overlap(x)

split_ldmap_range_overlap(x)
}
\arguments{
\item{x}{a (preferably sorted) ldmap_range vector (\code{length(x)} is referred to  as  \code{p})}
}
\value{
a sorted vector ldmap_ranges of length at least \code{p} and at most \verb{2p}(?) representing the same intervals

a sorted vector ldmap_ranges of length at least \code{p} and at most \verb{2p}(?) representing the same intervals
}
\description{
Take a vector of (preferably) sorted and (possibly) overlapping ldmap_ranges and create a new range of (sorted) non-overlapping ldmap_ranges

Take a vector of (preferably) sorted and (possibly) overlapping ldmap_ranges and create a new range of (sorted) non-overlapping ldmap_ranges
}
