% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_ldmap_region_overlap}
\alias{split_ldmap_region_overlap}
\title{Take a vector of (preferably) sorted and (possibly) overlapping ldmap_regions and create a new range of (sorted) non-overlapping ldmap_regions}
\usage{
split_ldmap_region_overlap(x)
}
\arguments{
\item{x}{a (preferably sorted) ldmap_region vector (\code{length(x)} is referred to  as  \code{p})}
}
\value{
a sorted vector ldmap_regions of length at least \code{p} and at most \verb{2p}(?) representing the same intervals
}
\description{
Take a vector of (preferably) sorted and (possibly) overlapping ldmap_regions and create a new range of (sorted) non-overlapping ldmap_regions
}
