% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{starts<-}
\alias{starts<-}
\title{Change start value for ldmap_region}
\usage{
starts(x) <- value
}
\arguments{
\item{x}{ldmap_region}

\item{value}{new  start}
}
\value{
an ldmap_region with all starts at \code{value}
}
\description{
Change start value for ldmap_region
}
