% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{starts}
\alias{starts}
\title{get starting position from a ldmap_region}
\usage{
starts(ldmap_region)
}
\arguments{
\item{ldmap_region}{the vector of ldmap_regions}
}
\description{
get starting position from a ldmap_region
}
