% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsnp.R
\name{subset_rds}
\alias{subset_rds}
\title{subset bigsnp by an ldmap_region}
\usage{
subset_rds(ldmr, reference_files, pattern = "sub")
}
\arguments{
\item{ldmr}{an ldmap_region}

\item{reference_files}{bigsnpr rds files}
}
\value{
a new rds file
}
\description{
subset bigsnp by an ldmap_region
}
\author{
Nicholas Knoblauch
}
