% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{translate_ldmr}
\alias{translate_ldmr}
\title{translate a vector of ldmap_regions by a constant factor}
\usage{
translate_ldmr(x, offset = 0L)
}
\arguments{
\item{x}{vector of ldmap_region}

\item{offset}{amount to move begin and end (recycled as necessary)}
}
\value{
vector of ldmap_regions
}
\description{
translate a vector of ldmap_regions by a constant factor
}
\examples{
# move all regions over by 100 base pairs
ldetect_EUR_100 <- translate_ldmr(ldetect_EUR,100L)
#move first region by 100 and second by 0
ld2 <- translate_ldmr(ldetect_EUR[1:2],c(100L,0L))
}
