% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{vec_proxy_compare.ldmap_snp}
\alias{vec_proxy_compare.ldmap_snp}
\title{Comparison method}
\usage{
\method{vec_proxy_compare}{ldmap_snp}(x, ...)
}
\arguments{
\item{x}{a vector of ldmap_snps}

\item{...}{unused}
}
\value{
a ranking of ldmap_snps
}
\description{
Comparison method
}
