% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{which_ldetect_window}
\alias{which_ldetect_window}
\title{assign ldmap snp or region to an ldetect block}
\usage{
which_ldetect_window(x, pop = "EUR", overlaps = FALSE)
}
\arguments{
\item{x}{vector of type ldmap_snp or ldmap_region}

\item{pop}{population for ldetect}

\item{overlaps}{logical indicating whether overlap is allowed (TRUE),
or only complete membership}
}
\value{
vector of lengtth length(x) with ldetect windows for each variant
}
\description{
assign ldmap snp or region to an ldetect block
}
