% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{window_ldmap_range}
\alias{window_ldmap_range}
\title{Create overlapping regions based on monotonic, point-level annotation}
\usage{
window_ldmap_range(ldmap_snp, cm, window = 1)

window_ldmap_range(ldmap_snp, cm, window = 1)
}
\arguments{
\item{ldmap_snp}{a (sorted) ldmap_snp vector (\code{length(ldmap_snp)} is referred to  as  \code{p})}

\item{cm}{a numeric vector of (length \code{p}) per-snp annotations (e.g cumulative recombination rate)}

\item{window}{the window width.}
}
\value{
a vector of \code{ldmap_range}s of length \code{p} giving the window for each SNP.  The width of the window
is defined for a target snp \code{ldmap_snp[i]}, as having the chromosome
from \code{ldmap_snp[i]} and including the position of all \code{ldmap_snp[j]} snps such that \code{abs(cm[i]-cm[j])<window} for all values of \code{j}

a vector of \code{ldmap_range}s of length \code{p} giving the window for each SNP.  The width of the window
is defined for a target snp \code{ldmap_snp[i]}, as having the chromosome
from \code{ldmap_snp[i]} and including the position of all \code{ldmap_snp[j]} snps such that \code{abs(cm[i]-cm[j])<window} for all values of \code{j}
}
\description{
Create overlapping regions based on monotonic, point-level annotation

Create overlapping regions based on monotonic, point-level annotation
}
