% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{write_bed}
\alias{write_bed}
\title{Write a dataframe with an ldmap_region column as a bed file}
\usage{
write_bed(
  df,
  path,
  region_col = region_cols(df),
  write_fun = readr::write_tsv,
  ...
)
}
\arguments{
\item{df}{dataframe with at least one ldmap_region column}

\item{path}{output file to write to}

\item{region_col}{name of ldmap_region column}

\item{write_fun}{function to use for writing bed file (defautls to \code{readr}'s \code{write_tsv})}

\item{...}{arguments passed to \code{write_fun}}
}
\value{

}
\description{
Write a dataframe with an ldmap_region column as a bed file
}
\examples{

temp_df <- tibble::tibble(ld = ldetect_EUR)
pth <- fs::file_temp(ext = "bed")
write_bed(temp_df, pth)
}
