% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{write_plink}
\alias{write_plink}
\title{write trio of bim,bed and fam files to plink format}
\usage{
write_plink(bim_df, bed_l = bim_df$genotypes, fam_df, file)
}
\arguments{
\item{bim_df}{dataframe with bim info}

\item{bed_l}{list of ldmap_gt vectors}

\item{fam_df}{data for individuals}

\item{file}{file to write (can contain \verb{.bed,.bim or .fam} prefix)}
}
\description{
write trio of bim,bed and fam files to plink format
}
