% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{write_plink_bim}
\alias{write_plink_bim}
\title{Write snp dataframe as plink bim file}
\usage{
write_plink_bim(
  bim_df,
  path,
  snp_col = snp_cols(bim_df),
  id_col = "rsid",
  map_col = "map",
  append = FALSE
)
}
\arguments{
\item{bim_df}{dataframe (as from read_bim) extra columns are ignored}

\item{path}{output file (or prefix)}

\item{snp_col}{which column should be expanded (contains ldmap_snp)}

\item{id_col}{which column has snp names (rsids)}

\item{map_col}{which column has genetic map values}
}
\description{
Write snp dataframe as plink bim file
}
