% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{write_plink_fam}
\alias{write_plink_fam}
\title{write plink fam data}
\usage{
write_plink_fam(fam_df, file)
}
\arguments{
\item{fam_df}{fam dataframe}

\item{file}{path to write}
}
\description{
write plink fam data
}
