% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsnp.R
\name{align_reference}
\alias{align_reference}
\title{Align to bigsnp reference}
\usage{
align_reference(gwas_df, reference_file, remove_missing = TRUE)
}
\arguments{
\item{gwas_df}{gwas summary stats}

\item{reference_file}{bigsnp reference file}
}
\value{
gwas summary statistics (ready for fine-mapping etc)
}
\description{
Align to bigsnp reference
}
