% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alt_alleles}
\alias{alt_alleles}
\title{get ref alleles from a ldmap_snp}
\usage{
alt_alleles(struct_vec, as_ascii_int = TRUE)
}
\arguments{
\item{struct_vec}{the vector of SNPs}
}
\description{
get ref alleles from a ldmap_snp
}
