% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_integer_ldmap_allele}
\alias{as_integer_ldmap_allele}
\title{coerce ldmap allele to integer}
\usage{
as_integer_ldmap_allele(x)
}
\arguments{
\item{x}{ldmap_allele vec}
}
\description{
coerce ldmap allele to integer
}
