% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{chromosomes<-}
\alias{chromosomes<-}
\title{Change chromosome value for a range or SNP}
\usage{
chromosomes(x) <- value
}
\arguments{
\item{x}{ldmap_range or ldmap_snp}

\item{value}{new chromosome value}
}
\value{
an ldmap_range or ldmap_snp with the appropriate chromsome
}
\description{
Change chromosome value for a range or SNP
}
