% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{convex_hull}
\alias{convex_hull}
\title{Find convex hull of vector of ranges (or SNPs )}
\usage{
convex_hull(x)

convex_hull(x)
}
\arguments{
\item{vector}{of ldmap_range or ldmap_snp of}
}
\value{
an ldmap range containg all the ranges (or snps)

an ldmap range containg all the ranges (or snps)
}
\description{
Find convex hull of vector of ranges (or SNPs )

Find convex hull of vector of ranges (or SNPs )
}
