% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{ends<-}
\alias{ends<-}
\title{Change end value for ldmap_region}
\usage{
ends(x) <- value
}
\arguments{
\item{x}{ldmap_region}

\item{value}{new end}
}
\value{
an ldmap_region with all starts at \code{value}
}
\description{
Change end value for ldmap_region
}
