% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ends}
\alias{ends}
\title{get end position from a ldmap_region}
\usage{
ends(ldmap_region)
}
\arguments{
\item{ldmap_region}{the vector of ldmap_regions}
}
\description{
get end position from a ldmap_region
}
