% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{ends}
\alias{ends}
\title{get end position from a ldmap_range}
\usage{
ends(ldmap_range)

ends(ldmap_range)
}
\arguments{
\item{ldmap_range}{the vector of ldmap_ranges}
}
\description{
get end position from a ldmap_range

get end position from a ldmap_range
}
