% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{explode_snp_struct}
\alias{explode_snp_struct}
\title{convert snp struct back into seperate columns}
\usage{
explode_snp_struct(
  df,
  ldmap_snp = snp_cols(df),
  alleles_to_character = FALSE,
  remove = TRUE
)
}
\arguments{
\item{df}{dataframe with at least one ldmap_snp_struct}

\item{ldmap_snp}{}
}
\value{

}
\description{
convert snp struct back into seperate columns
}
\examples{
#TODO

}
