% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.R
\name{format.ldmap_allele}
\alias{format.ldmap_allele}
\title{Formatting method for ldmap allele}
\usage{
\method{format}{ldmap_allele}(x, ...)
}
\arguments{
\item{x}{ldmap_allele}

\item{...}{unused}
}
\value{

}
\description{
Formatting method for ldmap allele
}
