% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_strand_ambiguous}
\alias{is_strand_ambiguous}
\title{Which SNPs are strand ambiguous}
\usage{
is_strand_ambiguous(x)
}
\arguments{
\item{x}{the ldmap_snp struct}
}
\description{
Which SNPs are strand ambiguous
}
