% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldmap_snp_2_dataframe}
\alias{ldmap_snp_2_dataframe}
\title{Convert ldmap snp back to dataframe}
\usage{
ldmap_snp_2_dataframe(ldmap_snp, alleles_to_character = FALSE)
}
\arguments{
\item{ldmap_snp}{the vector of SNPs}
}
\description{
Convert ldmap snp back to dataframe
}
