% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{migrate_ldmap_snp}
\alias{migrate_ldmap_snp}
\title{migrate from old representation of ldmap_snp to new representation}
\usage{
migrate_ldmap_snp(x)
}
\arguments{
\item{x}{ldmap_snp vec}
}
\description{
migrate from old representation of ldmap_snp to new representation
}
