% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nearest_snp_range}
\alias{nearest_snp_range}
\title{Assign ranges to nearest ranges}
\usage{
nearest_snp_range(query, target)
}
\arguments{
\item{query}{vector of query ldmap_snps}

\item{target}{vector of target ldmap_snps (must be sorted)}
}
\value{
a vector of integers of length \code{length(ldmap_range_query)} with the index of the \code{ldmap_range_target}
}
\description{
Assign ranges to nearest ranges
}
