% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_ldmap_snp}
\alias{new_ldmap_snp}
\title{Creation of new ldmap_snps}
\usage{
new_ldmap_snp(
  chrom = as.integer(c()),
  pos = as.numeric(c()),
  ascii_ref = as.integer(c()),
  ascii_alt = as.integer(c())
)
}
\arguments{
\item{chrom}{an integer vector of chromosomes}

\item{pos}{a double vector of positions}

\item{ascii_ref}{an optional integer vector of alternate allele (see \code{?utf8ToInt})}
}
\description{
Creation of new ldmap_snps
}
