% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{old_ldmap_snp}
\alias{old_ldmap_snp}
\title{migrate from new representation of ldmap_snp to old representation}
\usage{
old_ldmap_snp(x)
}
\arguments{
\item{x}{ldmap_snp vec}
}
\description{
migrate from new representation of ldmap_snp to old representation
}
