% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{overlaps}
\alias{overlaps}
\title{Check for overlap}
\usage{
overlaps(x, y)
}
\arguments{
\item{x}{vector of ldmap_snp or ldmap_region}

\item{y}{vector of ldmap_snp or ldmap_region}
}
\value{
boolean vector
}
\description{
Check for overlap
}
