% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap_utils.R
\name{read_hap}
\alias{read_hap}
\title{read \code{.hap} (or \code{.hap.gz}) files into a list of ldmap_ht}
\usage{
read_hap(x, ...)
}
\arguments{
\item{x}{path to \code{.hap} (or \code{.hap.gz}) file}
}
\value{
list_of ldmap_ht
}
\description{
read \code{.hap} (or \code{.hap.gz}) files into a list of ldmap_ht
}
