% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{read_plink_bed}
\alias{read_plink_bed}
\title{read plink bed file}
\usage{
read_plink_bed(f, N = NA, p = NA, subset = NULL)
}
\arguments{
\item{f}{path to bed file (or prefix)}

\item{N}{number of samples (optional, will read the fam file to figure it out)}

\item{p}{number of variants (optional will read the bim file to figure it out)}
}
\value{
list_of ldmap_gt
}
\description{
read plink bed file
}
