% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_utils.R
\name{read_snp_region_h5}
\alias{read_snp_region_h5}
\title{Read HDF5 file with snp info}
\usage{
read_snp_region_h5(h5file, ldmr, datapath = "snp", ...)
}
\arguments{
\item{h5file}{}

\item{ldmr}{ldmap_region scalar or vector}

\item{datapath}{datapath contain snp dataframe}

\item{...}{other arguments passed to read_df_h5

Requires \code{EigenH5}}
}
\value{
a dataframe with only SNPs from inside \code{ldmr}
}
\description{
Read HDF5 file with snp info
}
