% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (dellxps's conflicted copy
%   2020-01-06).R, R/RcppExports.R
\name{snp_in_range}
\alias{snp_in_range}
\title{Assign SNPs to ranges}
\usage{
snp_in_range(ldmap_snp, ldmap_range)

snp_in_range(ldmap_snp, ldmap_range)
}
\arguments{
\item{ldmap_snp}{vector of ldmap_snps (must be sorted)}

\item{ldmap_range}{vector of non-overlapping ldmap_ranges (must be sorted)}
}
\value{
a vector of integers of length \code{length(ldmap_snp)} with the index of the \code{ldmap_range}

a vector of integers of length \code{length(ldmap_snp)} with the index of the \code{ldmap_range}
}
\description{
Assign SNPs to ranges

Assign SNPs to ranges
}
