% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{snp_overlap_snp}
\alias{snp_overlap_snp}
\title{Assign SNPs to ranges}
\usage{
snp_overlap_snp(x, y)
}
\arguments{
\item{x}{vector of query ldmap_snps}

\item{y}{vector of target ldmap_snps}
}
\value{
a vector of integers of length \code{length(ldmap_snp)} with the index of the \code{ldmap_region}
}
\description{
Assign SNPs to ranges
}
