% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_utils.R
\name{starts<-}
\alias{starts<-}
\title{Change start value for ldmap_range}
\usage{
starts(x) <- value
}
\arguments{
\item{x}{ldmap_range}

\item{value}{new  start}
}
\value{
an ldmap_range with all starts at \code{value}
}
\description{
Change start value for ldmap_range
}
