% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{starts}
\alias{starts}
\title{get starting position from a ldmap_range}
\usage{
starts(ldmap_range)
}
\arguments{
\item{ldmap_range}{the vector of ldmap_ranges}
}
\description{
get starting position from a ldmap_range
}
